2015-9-29

+++ DataCenter/src/main/java/com/mamitang/dao/UserEntityMapper.java
@@ -26,5 +26,5 @@ public interface UserEntityMapper {
 
     List<UserEntity> getAllUsers(Map map);
 
-    Long getAllUsersCount();
+    int getAllUsersCount();
 }
